lab1/prediction_service/requirements.txt
Flask
joblib
scikit-learn
pandas
numpy
nibabel
gunicorn
google-cloud-storage 